Move protocol state initializer to initializer package
